Updated profile `SkillTypes` property to use string to enum converter Updated surrounding properties to not be null Updated related code to reflect nullability change

+++ Libraries/SPTarkov.Server.Core/Controllers/InventoryController.cs
@@ -192,7 +192,7 @@ public class InventoryController(
                         continue;
                     }
 
-                    profileSkill.Progress = mailEvent.Value;
+                    profileSkill.Progress = (double)mailEvent.Value;
                     logger.Success($"Set profile skill: {mailEvent.Entity} to: {mailEvent.Value}");
                     break;
                 }
